Fixed some errors in MCMC output
*There was a mistake in the paths of the np.savetxt used at the end of
the MCMC run (the path for the file containing the Period distributions
was being used for the MCMC results, and that of the MCMC results for
the period distribution!).

* There was another mistake in the for loop that runs the MCMC for each
transit. In particular, I had created three lists (one for the first
fit results of the transit, another one for its period distribution,
and a final one for its MCMC results) *inside* the for loop, which
means they were being converted into empty lists every time I was going
through the loop. They’re now outside and they work fine!

* the medfilter is currently deactivated.