* Added SNR calculation to identify missing transits and distinguish between good and bad light curves. Stars which have either a missing transit or a bad light curve (or both) are stored in a dictionary called “ugly_stars”. This is still a work in progress.
* Spent lots of hours trying to understand why the hell the first
optimisation with DFM’s Transit routine was not working. After trying
many different things, now it seems to work *fine* (not great, just
fine). The most important change is that i’ve set the initial guess of
the impact parameter to 0.001, which I believe really helps the code.
I’ve also removed the maxiter in the “fit_lc” function, even if
sometimes the optimizator does not manage to converge.

* Added a “maxDepth” for the prior of the depth in the “lnprior”
function, which is the one that the optimisation involving DFM’s
transit routine uses.

* Changed the prior limits of the depth from 0 (included) to maxDepth.